build(nx): update to latest dynamic config
