+++ package.json
@@ -33,7 +33,10 @@
     "@graphql-tools/apollo-engine-loader": "^8.0.0",
     "@ianwremmel/eslint-plugin-ianwremmel": "^6.2.2",
     "@jest/types": "^29.6.3",
+    "@nx/plugin": "^17.2.8",
     "@semrel-extra/npm": "^1.2.2",
+    "@swc-node/register": "^1.6.8",
+    "@swc/core": "^1.3.106",
     "@types/jest": "^29.5.10",
     "@types/lodash": "^4.14.200",
     "@types/node": "^20.0.0",
