+++ packages/@code-like-a-carpenter/nx-auto/src/index.js
@@ -1,7 +1,5 @@
 'use strict';
 
-module.exports = require('./nx-auto');
-
 if (require.main === module) {
   require('./cli');
 }
