+++ packages/@code-like-a-carpenter/nx/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@code-like-a-carpenter/nx",
+  "main": "./src/index.ts",
+  "dependencies": {
+    "@nx/devkit": "17.2.8",
+    "glob": "^10.3.10"
+  },
+  "private": true
+}
