+++ patches/nx+17.2.8.patch
@@ -1,0 +1,18 @@
+diff --git a/node_modules/nx/src/native/index.d.ts b/node_modules/nx/src/native/index.d.ts
+index 65c19a0..9587a75 100644
+--- a/node_modules/nx/src/native/index.d.ts
++++ b/node_modules/nx/src/native/index.d.ts
+@@ -1,6 +1,13 @@
+ /* tslint:disable */
+ /* eslint-disable */
+ 
++type JsInputs = unknown
++type JsExternal = unknown
++type HashInstruction = unknown
++type ProjectFiles = unknown
++type ProjectRootMappings = unknown
++type NapiDashMap = unknown
++
+ /* auto-generated by NAPI-RS */
+ 
+ export class ExternalObject<T> {
